Move schedule tests
