+++ src/domain/flow-system/Cargo.toml
@@ -36,3 +36,6 @@ url = { version = "2", default-features = false }
 # TODO: Make serde optional
 serde = { version = "1", default-features = false, features = ["derive"] }
 serde_with = { version = "3", default-features = false }
+
+[dev-dependencies]
+kamu = { workspace = true }
